Fix CBOR fields showing "ld" in some device logs.
The <inttypes.h> typedef PRId64 is defined as the format specifier "%lld" (for some platforms), however this is unimplemented on some devices. This causes it to evaluate to the characters "ld" when used as a printf format specifier.

+++ api/oc_rep.c
@@ -1034,7 +1034,7 @@ oc_rep_to_json_format(oc_rep_t *rep, char *buf, size_t buf_size, int tab_depth,
     }
     case OC_REP_INT: {
       num_char_printed =
-        snprintf(buf, buf_size, "%" PRId64, rep->value.integer);
+        snprintf(buf, buf_size, "%d", (int)rep->value.integer);
       OC_JSON_UPDATE_BUFFER_AND_TOTAL;
       break;
     }
@@ -1097,7 +1097,7 @@ oc_rep_to_json_format(oc_rep_t *rep, char *buf, size_t buf_size, int tab_depth,
         oc_rep_i_get_int_array(rep, rep->iname, &int_array, &int_array_size);
       }
       for (size_t i = 0; i < int_array_size; i++) {
-        num_char_printed = snprintf(buf, buf_size, "%" PRId64, int_array[i]);
+        num_char_printed = snprintf(buf, buf_size, "%d", (int)int_array[i]);
         OC_JSON_UPDATE_BUFFER_AND_TOTAL;
         if (i < int_array_size - 1) {
           num_char_printed = (pretty_print) ? snprintf(buf, buf_size, ", ")
